backend/requirements.txt
fastapi
uvicorn
confluent-kafka
pydantic
python-dotenv
kafka-python
docker
